### 2019-05-18 #4 @marskar
Nbless: (de)construct, convert, execute, and prepare slides from Jupyter notebooks.
Submitting Author: Name (@marskar)
Repository Link: https://github.com/marskar/nbless
Documentation Link: https://marskar.github.io/nbless/
PyPI Link: https://pypi.org/project/nbless/

---

```
The Nbless Python package allows you to
- (de)construct,
- convert,
- execute, and
- prepare slides from Jupyter notebooks.

The Nbless Python package consists of 6 Python functions and shell commands:

- nbconv, which converts a notebook into various formats.
- nbdeck, which prepares a notebook to be viewed as or converted into slides.
- nbexec, which runs a notebook from top to bottom and saves an executed version.
- nbless, which calls nbuild and nbexec to create and execute a notebook.
- nbraze, which extracts code and markdown files from a notebook.
- nbuild, which creates a notebook from code and markdown files.
```

- Please indicate which [category or categories](https://pyopensci.github.io/dev_guide/peer_review/peer_review_proc.html#aims-and-scope) this package falls under:

Jupyter notebooks are important for reproducibility, because they facilitate literate programming, the combination of code and text. This way, descriptions of methods and results can be adjacent to the code used to obtain the results.
I frequently use Jupyter notebooks in my work and in the Python classes I teach, but access may be limited due to the underlying JSON structure of notebooks.
With Nbless, anyone can create Jupyter notebooks from simple source files or extract code and markdown files from Jupyter notebooks. This can promote the modularity principle without sacrificing the benefits of literate programming or the convenience of a single file with all code, text, and output for the project.
Code and markdown files can be easily be accessed with any text editor or even opened in a browser. Importantly, the original Jupyter notebook can be recreated from code and markdown files.

The target audience is anyone who uses Jupyter notebooks. This package may help teachers and researchers to improve the efficiency of their Jupyter notebook workflows. One of the major features of Nbless is that it can facilitate the creation of slides from Jupyter notebooks.  I can imagine teachers and researchers writing markdown and code files and then using `nbless` to creating Jupyter notebooks to share with their students and colleagues.

Related packages include:

- [`pandoc`](https://pandoc.org/MANUAL.html#creating-jupyter-notebooks-with-pandoc)
- [`jupytext`](https://github.com/mwouts/jupytext)
- [`notedown`](https://github.com/aaren/notedown)

These packages can all convert Jupyter notebooks to other formats: markdown files (all three) or Python scripts (`jupytext`). Nbless can take a more modular approach to file conversion by extracting the contents of each notebook cell into a separate file (cell -> file) or using a source file to create each notebook cell (file -> cell). Looking beyond simple file conversion, Nbless includes a tool for making slides from notebooks (by setting `slide_type` in notebook metadata).

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*
