### 2019-05-20 #4 @lwasser
thank you @kysolvik  :) @marskar welcome to pyopensci!! i'd like @choldgraf to look at your three submissions given his is on the Jupyter dev team!!
