### 2019-05-20 #4 @marskar
Hi @kysolvik, @lwasser, and @choldgraf,
Thanks for the quick follow-up.
No rush with the review. I just want to see if any of these three packages would be of interest to pyopensci. @choldgraf, I do not have a preference, please feel free to prioritize or deprioritize as you see fit.
- Nbless is related to Jupyter. It uses `nbformat` and `nbconvert`. I wrote it after watching Joel Grus' [I don't like notebooks talk](https://youtu.be/7jiPeIFXb6U) 😃
- Rmdawn is similar to Nbless, except its for R markdown and R files.
- Gitone is a package that I use all the time. You can think of it as a `.gitconfig` in the form of a python package 😃It will be relevant to everyone who uses or wants to use git on the command line.
