### 2019-05-20 #4 @lwasser
Thank you so much @marskar we  appreciate your engaging with us and all of the submissions! we will have a look at all three individually and will then get back to you! because we are new and still building our infrastructure, we may need to space things out based upon resources and time. However we will get back to you and are  appreciative of the submissions!

By the way, are you interested in being a package reviewer in the future?
