### 2019-05-20 #4 @marskar
Thanks @lwasser!
Again, no rush with the reviews.
I understand that pyopensci is new. I was following ropensci for a while and was excited to hear about pyopensci! I use R a lot, but I prefer to develop in Python (for many reasons!).
I would be honored to be a pyopensci package reviewer in the future!
