### 2019-05-20 #4 @lwasser
oh wonderful. thank you @marskar !! You are also welcome to drop in on one of our meetings (we meet every three weeks) via zoom. if you are interested please send me an email and i'll add you to the invite list. we look forward to working with you in the future. We will need reviewers so it's wonderful that you are willing to contribute in that way!

More on your submissions soon!! :)
