### 2019-05-22 83d8aac
add next steps to README.rst
