### 2019-05-26 41c6762
add tests and enable nbuild, nbless, and nbconv to accept pathlib Path objects
