+++ src/nbless/nbconv.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 from pathlib import Path
-from typing import Tuple
+from typing import Tuple, Union
 
 from nbconvert.exporters.base import get_exporter
 
 
-def nbconv(filename: str, exporter: str = "python") -> Tuple[str, str]:
+def nbconv(filename: Union[str, Path], exporter: str = "python") -> Tuple[str, str]:
     """Convert a notebook into various formats using ``nbformat`` exporters.
 
     :param filename: The name of the input notebook file.
@@ -15,6 +15,8 @@ def nbconv(filename: str, exporter: str = "python") -> Tuple[str, str]:
            pdf requires latex, 'notebook' does nothing,
            slides need to served (not self-contained).
     """
-    contents, resources = get_exporter(exporter)().from_filename(filename)
+    contents, resources = get_exporter(exporter)().from_filename(
+            filename.name if isinstance(filename, Path) else filename
+            )
     out_name = Path(filename).stem + resources.get("output_extension", ".txt")
     return out_name, contents
