### 2019-05-26 72fe4e4
add tests and enable nbuild, nbless, and nbconv to accept pathlib Path objects
