+++ src/nbless/nbdeck.py
@@ -2,7 +2,7 @@
 import nbformat
 
 
-def nbdeck(filename: str) -> nbformat.notebooknode.NotebookNode:
+def nbdeck(in_file: str) -> nbformat.notebooknode.NotebookNode:
     """ Set up a Jupyter notebook to be viewed as or converted into slides.
 
     Sets `slide_type` to ``slide`` for markdown cells that start with headers.
@@ -10,9 +10,9 @@ def nbdeck(filename: str) -> nbformat.notebooknode.NotebookNode:
     - view Jupyter notebooks on nbviewer or with the RISE extension
     - create an HTML slideshow with ``nbconv`` or ``jupyter nbconvert``.
 
-    :param filename: The name of the input Jupyter notebook file.
+    :param in_file: The name of the input Jupyter notebook file.
     """
-    nb = nbformat.read(filename, as_version=4)
+    nb = nbformat.read(in_file, as_version=4)
     for n, cell in enumerate(nb.cells):
         if cell.cell_type == "markdown" and cell.source.startswith("#"):
             nb.cells[n].metadata = {"slideshow": {"slide_type": "slide"}}
