+++ src/nbless/nbexec.py
@@ -6,13 +6,13 @@ from nbconvert.preprocessors import ExecutePreprocessor
 from nbformat.notebooknode import NotebookNode
 
 
-def nbexec(filename: str, kernel: str = "python3") -> Tuple[str, NotebookNode]:
+def nbexec(in_file: str, kernel: str = "python3") -> Tuple[str, NotebookNode]:
     """Create an executed notebook without modifying the input notebook.
 
-    :param filename: The name of the Jupyter notebook file to be executed.
+    :param in_file: The name of the Jupyter notebook file to be executed.
     :param kernel: The programming language used to execute the notebook.
     """
-    nb = nbformat.read(filename, as_version=4)
+    nb = nbformat.read(in_file, as_version=4)
     ep = ExecutePreprocessor(timeout=600, kernel_name=kernel)
     ep.preprocess(nb, {"metadata": {"path": "."}})
     return nb
