+++ requirements.txt
@@ -1,3 +1,7 @@
 -e .
 jupyter==1.0.0
 click==7.0
+pytest
+pytest-mypy
+matplotlib
+numpy
