### 2019-05-26 ca86f1e
Pin pytest to latest version 4.5.0
