### 2019-05-26 4dcdbd3
Pin pytest-mypy to latest version 0.3.2
