### 2019-05-26 9316b07
Pin matplotlib to latest version 3.1.0
