### 2019-05-26 61e11de
Pin numpy to latest version 1.16.3
