+++ requirements.txt
@@ -4,4 +4,4 @@ click==7.0
 pytest
 pytest-mypy
 matplotlib
-numpy
+numpy==1.16.3
