### 2019-05-29 4f5ba8d
Pin matplotlib to latest version 3.1.0
