+++ requirements.txt
@@ -3,5 +3,5 @@ jupyter==1.0.0
 click==7.0
 pytest
 pytest-mypy
-matplotlib
+matplotlib==3.1.0
 numpy==1.16.4
