### 2019-05-29 3a53b38
Pin pytest-mypy to latest version 0.3.2
