### 2019-05-29 0c12344
Pin pytest to latest version 4.5.0
