+++ requirements.txt
@@ -1,7 +1,7 @@
 -e .
 jupyter==1.0.0
 click==7.0
-pytest
+pytest==4.5.0
 pytest-mypy==0.3.2
 matplotlib
 numpy==1.16.4
