+++ README.rst
@@ -3,10 +3,6 @@ Nbless: a Python package for programmatic Jupyter notebook workflows
 
 |Status| |PyPI| |Updates|
 
-.. image:: https://www.repostatus.org/badges/latest/active.svg
-   :alt: Project Status: Active – The project has reached a stable, usable state and is being actively developed.
-   :target: https://www.repostatus.org/#active
-
 Introduction
 ------------
 
