### 2019-05-29 22a483e
add .travis.yml and edit badges in readme
