+++ tox.ini
@@ -3,21 +3,19 @@ envlist = py27, py34, py35, py36, flake8
 
 [travis]
 python =
+    3.7: py37
     3.6: py36
-    3.5: py35
-    3.4: py34
-    2.7: py27
 
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 rmdawn
+commands = flake8 nbless
 
 [testenv]
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
-    -r{toxinidir}/requirements_dev.txt
+    -r{toxinidir}/requirements.txt
 ; If you want to make tox run the tests with the same versions, create a
 ; requirements.txt with the pinned versions and uncomment the following line:
 ;     -r{toxinidir}/requirements.txt
