+++ input_files/letters.R
@@ -1,2 +1,0 @@
-letters[1:26]
-LETTERS[1:26]
