+++ input_files/index.md
@@ -1,184 +1,0 @@
-# The `nbless` python package
-
-Using `nbless` you can create and execute [Jupyter Notebooks](http://jupyter-notebook.readthedocs.io/en/latest/examples/Notebook/What%20is%20the%20Jupyter%20Notebook.html) in
-- your terminal or
-- your favorite Python environment (e.g. [PyCharm](https://www.jetbrains.com/pycharm/) or [Visual Studio Code](https://code.visualstudio.com/docs/python/python-tutorial)).
-
-The `nbless` python package consists of 5 functions:
-- `nbuild`, which creates a notebook from python scripts and plain text files, e.g. markdown (`.md`) and text (`.txt`) files.
-- `nbexec`, which runs a notebook from top to bottom and saves an executed version, leaving the source notebook untouched.
-- `nbcode`, which converts a notebook into a code file. The extension is determined by the kernel. Markdown cells become comments.
-- `nbless`, which calls `nbuild`, `nbcode`, and `nbexec` to create and execute a notebook.
-- `nbhtml`, which converts a notebook into an HTML document.
-
-All of the above function work for Python _and_ R code, with the caveat that nbexec and nbless require the kernel argument to be set to `ir` if R code files (`.R`) are included.
-
-If you want to execute a Jupyter notebook that contains both R and Python code, you will have to put the R code in Python scripts (`.py`) and use [rpy2](https://rpy2.readthedocs.io/) with the default kernel (`python3`).
-
-To run Rmd files that include R and Python code, you will need the `reticulate` R package.
-
-The `nb` functions rely on the `nbconvert` and `nbformat` modules that are included with `jupyter`. The `catrmd` function has no dependencies beyond Python.
-
-## Installation
-
-```sh
-pip install nbless
-```
-
-or clone the [repo](https://github.com/marskar/nbless), e.g. `git clone https://github.com/marskar/nbless` and either use locally, e.g. `python nbless.py README.md plot.py notes.txt` or install using setup.py, e.g. `python setup.py install`.
-
-## Basic usage: terminal
-
-### Creating a Jupyter notebook with `nbuild` in the terminal
-
-```sh
-nbuild README.md plot.py notes.txt
-```  
-
-The `nbuild` function takes the contents of Python code files (`.py`) stores them as [Jupyter Notebook code cells](https://jupyter-notebook.readthedocs.io/en/stable/examples/Notebook/Running%20Code.html). The contents of all other files are stored in [markdown cells](https://jupyter-notebook.readthedocs.io/en/stable/examples/Notebook/Working%20With%20Markdown%20Cells.html).
-
-The default output filename for `nbuild` is `unexecuted.ipynb`. By default, input and output files are located in the current directory (`'./'`).
-
-You can provide a more descriptive filename for the unexecuted notebook (`-u`) and set different input  (`-i`) and output  (`-o`) filepaths:
-
-```sh
-nbuild README.md plot.py notes.txt --input_path input_files --output_name raw.ipynb --output_path notebooks/
-# Or
-nbuild README.md plot.py notes.txt -i input_files -n raw.ipynb -o notebooks/
-```
-
-If you only want to execute a notebook, run `nbexec`.
-
-### Executing a notebook with `nbexec` in the terminal
-
-
-```sh
-nbexec raw.ipynb
-```
-
-The `nbexec` creates a copy of the input notebook, runs it from top to bottom and saves it. The default name of the first notebook is `unexecuted.ipynb` while the executed notebook is called `executed.ipynb` by default. The default filepath where the notebooks are saved is the current directory (`'./'`).
-
[…]
