+++ input_files/notes.txt
@@ -1,2 +1,0 @@
-cryzen codecademy
-PyDataDC
