+++ input_files/header.yml
@@ -1,5 +1,0 @@
----
-title: "Untitled"
-output: html_document
-output: html_notebook
----

[…]
