+++ test_nbless.py
@@ -1,4 +1,0 @@
-def test_nbless():
-    """Test slidoc"""
-    assert 3 + 2 == 5
-
