### 2019-05-29 6693cff
fix type hint error in conf.py and install pypandoc for tests
