### 2019-05-29 2c88544
add pandoc install and python version 3.7 to travis.yml
