+++ .travis.yml
@@ -1,7 +1,11 @@
 language: python
 python:
 - 3.6
-install: pip install -U tox-travis pytest-mypy jupyter pypandoc
+- 3.7
+install:
+  - pip install -U tox-travis pytest-mypy jupyter
+  - sudo apt-get install pandoc
+
 script: tox
 deploy:
   provider: pypi
@@ -12,4 +16,4 @@ deploy:
   on:
     tags: true
     repo: marskar/nbless
-    python: 3.6
+    python: 3.7
