### 2019-05-29 5c4ad4a
add python version 3.7 to tox.ini
