+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27, py34, py35, py36, flake8
+envlist = py36, py37, flake8
 
 [travis]
 python =
