### 2019-05-29 177cd33
remove pathlib support from nbuild
