### 2019-05-29 2fe2a87
fix type hint error in nbuild
