+++ src/nbless/nbuild.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
-from typing import List, Union
+from typing import List
 from pathlib import Path
 
 from nbformat.v4 import new_notebook, new_code_cell, new_markdown_cell
 from nbformat.notebooknode import NotebookNode
 
 
-def nbuild(in_files: str) -> NotebookNode:
+def nbuild(in_files: List[str]) -> NotebookNode:
     """Create an unexecuted Jupyter notebook from markdown and code files.
 
     :param in_files: A list of source file names.
