### 2019-05-29 400bbe3
try to fix paths in tests
