+++ README.rst
@@ -1,7 +1,7 @@
 Nbless: a Python package for programmatic Jupyter notebook workflows
 ====================================================================
 
-|License| |PyPI| |Status| |Updates| |Versions|
+|Build| |License| |PyPI| |Status| |Updates| |Versions|
 
 Introduction
 ------------
@@ -374,6 +374,8 @@ Currently, notebook metadata is lost when using ``nbraze``/``nbuild``/``nbless``
 - Enable ``nbuild``/``nbless`` to accept metadata via a ``metadata.json`` file.
 - Enable ``nbraze`` to output metadata via a ``metadata.json`` file.
 
+.. |Build| image:: https://travis-ci.org/marskar/nbless.svg?branch=master
+   :target: https://travis-ci.org/marskar/nbless
 .. |License| image:: https://img.shields.io/badge/License-MIT-brightgreen.svg
    :target: https://opensource.org/licenses/MIT
 .. |PyPI| image:: https://img.shields.io/pypi/v/nbless.svg
