### 2019-05-29 4227ca8
try to fix pandoc HOME error in tests
