+++ .travis.yml
@@ -3,9 +3,12 @@ language: python
 python:
 - 3.6
 - 3.7
+before_install:
+  - sudo apt-get update -qq
+  - sudo apt-get install -y pandoc
+  - export HOME=/home/travis
 install:
   - pip install -U tox-travis pytest-mypy jupyter
-  - sudo apt-get install pandoc
 
 script: tox
 deploy:
