### 2019-05-29 4ba360d
try to fix pandoc HOME error
