### 2019-05-29 f8d7c15
try to fix pandoc HOME error by building from source
