### 2019-05-29 2d55d95
try to fix pandoc HOME error by using addons
