+++ .travis.yml
@@ -3,17 +3,7 @@ language: python
 python:
 - 3.6
 - 3.7
-before_install:
-  - sudo apt-get update -qq
-  - sudo apt-get install -y pandoc
-  - export HOME=/home/travis
 install:
-  - wget https://hackage.haskell.org/package/pandoc-1.17.0.3/pandoc-1.17.0.3.tar.gz
-  - tar xvzf pandoc-1.17.0.3.tar.gz
-  - cd pandoc-1.17.0.3
-  - curl -sSL https://get.haskellstack.org/ | sh
-  - stack setup
-  - stack install
   - pip install -U tox-travis pytest-mypy jupyter
 script: tox
 deploy:
@@ -26,3 +16,6 @@ deploy:
     tags: true
     repo: marskar/nbless
     python: 3.7
+addons:
+  apt_packages:
+    - pandoc
