### 2019-05-29 4bb9ef6
give up on trying to fix pandoc HOME error
