### 2019-05-29 6335f7b
Bump version to 0.2.26
