### 2019-05-30 #4 @lwasser
ok @marskar thank you so much for this submission (and for the other 3!!). After our meeting discussion, let's plan to review this package first.

@leouieda and @ocefpaf will be the reviewers on this package. To begin the formal submission process, please submit a new issue that is an actual submission issue. Here you will be asked if this will be reviewed for JOSS as well as if you are OK with reviewers submitting issues to your repo directly.

Once you have submitted the review issue , I will

1.  formally assign @ocefpaf and @leouieda  as reviewers of this package
2. provide a review deadline and
3. ask you to close this issue given the new issue!

please let me know if you have any questions. looking forward to seeing this package go through the review process!
