### 2019-05-30 #4 @marskar
Thanks @lwasser!
I'm very excited that Nbless is going to be reviewed!
I made a new issue (#7) so I already closed the pre-submission inquiry issue (#4).
